% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generator-class.R
\name{BiocGenerator}
\alias{BiocGenerator}
\title{Construct, yield, and iterate over rectangular objects}
\usage{
BiocGenerator(source, dim = base::dim(source), yieldSize = dim(source))
}
\description{
Construct, yield, and iterate over rectangular objects
}
\examples{
se <- SummarizedExperiment(matrix(rnorm(1000), 10))
gen <- BiocGenerator(se, yieldSize = c(NA, 20))
}
